Refactoring: security utils class is added
